魔方密码V4.0.0.0

+++ MagicPwd/src/com/magicpwd/e/maoc/mfun/DeleteAction.java
@@ -17,6 +17,8 @@
 package com.magicpwd.e.maoc.mfun;
 
 import com.magicpwd.__a.maoc.AMaocAction;
+import com.magicpwd._cons.LangRes;
+import com.magicpwd._util.Lang;
 
 /**
  *
@@ -28,7 +30,10 @@ public class DeleteAction extends AMaocAction
     @Override
     public void actionPerformed(java.awt.event.ActionEvent e)
     {
-        maocPtn.deleteFun();
+        if (Lang.showFirm(maocPtn, LangRes.P30FBA0C, "确认要删除此函数吗？\n删除后其它依赖于此函数的常量或函数将无法参与运算！") == javax.swing.JOptionPane.YES_OPTION)
+        {
+            maocPtn.deleteFun();
+        }
     }
 
     @Override
